requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
kafka-python==2.0.2
openai==1.12.0
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
uvloop==0.19.0
asyncio-mqtt==0.16.1
redis==5.0.1
